requirements.txt
python-telegram-bot==13.*
python-dotenv==1.*
redis==4.5.*